Refactor app startup process and add video frame generation utility
- Updated start.cmd to launch Flask app with reload option.
- Introduced src/utils.py for generating video frames from a source.
- Enhanced style.css with new form styles and added horizontal flip effect for video.
- Replaced old settings.py with a new version utilizing dotenv for environment variables.
- Updated script.js to correct fetch URL for face capture and improved registration form.
- Added test.html for UI testing.
- Improved register.html with better form structure and styling.
